web-api-py-c++/requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Cython==0.29.32
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
h11==0.12.0
httpcore==0.15.0
httplib2==0.21.0
httpx==0.23.0
idna==3.4
oauthlib==3.2.2
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.0
python-telegram-bot==20.0a4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
snakeviz==2.1.1
sniffio==1.3.0
tornado==6.2
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12